+++ src/util/virfirewalld.h
@@ -29,7 +29,7 @@ typedef enum {
     VIR_FIREWALLD_BACKEND_LAST,
 } virFirewallDBackendType;
 
-int virFirewallDGetVersion(unsigned long *version);
+int virFirewallDGetVersion(unsigned long long *version);
 int virFirewallDGetBackend(void);
 int virFirewallDIsRegistered(void);
 int virFirewallDGetZones(char ***zones, size_t *nzones);
