+++ src/util/virstring.c
@@ -1022,7 +1022,7 @@ int virStringParseYesNo(const char *str, bool *result)
 
 /**
  * virStringParseVersion:
- * @version: unsigned long pointer to output the version number
+ * @version: unsigned long long pointer to output the version number
  * @str: const char pointer to the version string
  * @allowMissing: true to treat 3 like 3.0.0, false to error out on
  * missing minor or micro
@@ -1037,7 +1037,7 @@ int virStringParseYesNo(const char *str, bool *result)
  * Returns the 0 for success, -1 for error.
  */
 int
-virStringParseVersion(unsigned long *version,
+virStringParseVersion(unsigned long long *version,
                       const char *str,
                       bool allowMissing)
 {
